veyron2/storage/vstore: JoinAddressName -> Join
Unlike other services mounted on the mount table, the Veyron Store could not
be connected to using a Veyron name. The issue was that a leading '/' was
always prepended to the names upon trying to connect, possibly an artifact of
when we were only using endpoint addresses.

The issue is solved by using naming.Join(...) instead of
naming.JoinAddressName(...). The latter is the one that prepended the '/' and
prevented the MOUNTTABLE_ROOT environment variable from being correctly
prepended to Veyron names.

This change does not require any further code changes, but it does affect how
we use the programs. When connecting with an endpoint address, a leading '/'
must be used. Otherwise, the endpoint will be treated like any other name.
Connecting to the store with a Veyron name will now succeed.

Change-Id: I04d640a93f990dcc9bad1abd4e93b84cceb0d26c