veyron/runtimes/google: adding AppCycle service implementation to the runtime.
We create a new server for this purpose, and hook things up to the task tracker
mechanism. This CL also ensures that the task tracking channels get closed when
the runtime shuts down (to indicate when the shutdown is imminent).

To faciliate conveying the child's app cycle service name to the parent, we
introduce a config API and a config service (which subsumes the previous
callback mechanism into a more general API). The new mechanics are as follows:
- parent sends a config to the child, part of which is the parent's config
service veyron name;
- the child's app manager calls the parent's config service to let the parent
know what its own veyron name is;
- (if the child is a node manager), it calls the parent's config service to let
the parent know what the child's node manager service veyron name is.
The newly introduced config API and config service will be folded with the
config neighborhood service into "the" veyron config API.

Change-Id: I1ff427ae77097b012aef138df1e6347bf6623e4a