"veyron2/security": Added PublicIDStore interface
Each Veyron runtime is assumed to have exactly one Private key
however it may accumulate multiple PublicIDs over time by getting blessed
by others. Some of these PublicIDs may be deemed as useful for multiple
IPCs, and therefore must be stored (persisted).

This CL defines a PublicIDStore interface (in "veyron2/security") for
managing multiple PublicIDs and provides an implementation for it.
In a subsequent CL, I plan to embed a PublicIDStore inside a PrivateID
(and get rid of the embedded PublicID that we have at the moment).

Change-Id: I50d30c6643fd325c4d84a55590115e54a809ecd9