veyron2/storage/vstore/primitives: Implement storage.QueryStream.
Top level query results are pulled from the network as the client is
processing them.  Nested query results are pulled completely into memory
because the API allows the client to access them in an arbitrary order.
In the future, we will extend the network protocol to allow these
nested results to be sent via a separate rpc stream.  However, for
small nested result sets, it makes sense to inline them.  Therefore,
the code to do the inlining is valuable in the long run.

Change-Id: Idb16385acd4dbdd137dcf1f4912ee183a5f84f76