veyron2/services/mgmt/application,veyron2/services/mgmt/node: API changes around installing and updating applications.
The main change is to make the veyron name for the application to be downloaded
into an argument to Install (instead of being part of the receiver name).  The
second change is to introduce the concept of application title, which is
something that the app publisher provides, and in particular, we use to
construct the name suffix for the app installation. The idea is that all
app installations with the same title form a logical management group (so the
title acts like an app id of sorts). Updates cannot happen across title changes
(so the developer will have to do a brand new install in that case).

Change-Id: I923343553797d9016cdadaf0fc09884e0f21548d