veyron/services/store/memstore/query: Better handling of slashes in selection.
Ensure that selection of sub-sub-fields works.  Also, allow for aliases to
contain slashes.  I don't see any reason why they can't, and it matches the
behavior of selection of sub-sub-fields.

Fix a bug in selection: if an intermediate result does not have a value
for one of the sub pipelines, use nil as the value instead of aborting
the query (without an error message).

Change-Id: I631f7f80b562096f725ef566b648fe711534b94a