veyron2/storage/vstore/primitives: Remove bogus panic.
If one thread calls Cancel while another thread is in the
middle of iterating (i.e. Advance has already returned true,
but thread has not yet called Value), Value will panic.
The new behavior is that the iterating thread will finish
the current iteration and detect the cancel on the next
call to Advance.

Change-Id: I1dbd6a68075547e3078d40d4e5aef6cdd64310c9