veyron/security/auditor: Pacakge with interfaces for auditing sensitive actions.
The auditor package defines an interface for recording interesting events
and provides a security.PrivateID implementation that audits all operations
that use the private key.

The immediate motivation for this is the identityd HTTP server, where the
runtime's identity will use this auditing PrivateID to keep track of
all Veyron identities issued.

The same PrivateID implementation is also expected to be used by the
identity agent that ribrdb@ is working on.

Change-Id: Id7d0f1101a4b3217cece615830c76f81b1518aff