veyron2/vdl: Add Nilable types, which roughly correspond to pointers in Go. A value of type "?int64" can hold any int64, or may hold nil representing non-existence.  Unlike pointers there's no concept of sharing; although cyclic types can be represented, cyclic values cannot be represented.
Changed OneOf semantics so that its zero value is the zero value of the
0-indexed OneOf type, rather than nil.  The zero value of Any is still nil.

Changed the internal representation of values so that repSequence
represents Array, List and Struct.  This simplifies the code, and also lets
Lists take advantage of lazy-initialization.

This change only updates veyron2/vdl and veyron2/vdl/valconv.  The vdl
parser, compiler and codegen will be updated in a subsequent change.

Change-Id: Iffc91c9afa3d87c99644d6bf9ce3c080a405169b