veyron/vdl/parse: Add parser and compiler support for Nilable types.
The syntax for Nilable types uses question mark, and preceeds the base
type it is modifying.  Some examples:
  int64              - Non-nilable int64
  ?int64             - Nilable int64
  ?map[int64]?string - Nilable map, key int64, elem nilable string

The reason we use question mark ? rather than star * is to draw
attention to the fact that the semantics are different from pointers.
There's no notion of memory addresses or dereferencing, and thus
there's no concept of sharing.  I.e. while you can express cyclic
types in VDL, you can't express cyclic values.

The reason we prefix rather than suffix the symbol is because this is
how the rest of the grammar works.  In particular we can't use a
suffix, since there is ambiguity:
  map[int64]string? - What's nilable, the map or the string elem?

Note that the grammar allows syntax that we don't allow semantically.
E.g. we won't allow multi-nilable "??string", nor do we allow nilable
any "?any".  Those errors are caught in the compiler via
vdl.TypeBuilder.

We also disallow top-level nilable type definitions.  The purpose of
this rule is twofold:
1) Reduce confusion; the Nilable modifier cannot be hidden in a type
   definition, it must be explicitly mentioned on each use.
2) The Nilable concept is typically translated to pointers in
   generated languages, and many languages don't support named
   pointer types.

  type A string            // ok
  type B []?string         // ok
  type C struct{X ?string} // ok
  type C ?string           // bad
  type D ?struct{X string} // bad

Change-Id: I6907a10dea65c807c06212edc17214a662ef3383