veyron2/services/mgmt/stats: Add interface to access stats.
This new interface will be used to access stats for troubleshooting and
monitoring purposes. The stats objects are discoverable via the
Globbable interface and watchable via the GlobWatcher interface.

The type of the object values will be implementation specific, but
should be mostly numeric in nature, e.g. counters, memory usage, latency
metrics, etc.

Change-Id: Ia265adc71b2abc7428adaef1a92bb2cbdc8f25f2