veyron/runtimes/google/ipc: Add deadline and cancellation propagation to the context.
I have revised the CL and it's worth a good look now.  I have moved the derivation
functions to the interface instead of being at the top level as in go.nets Context.
This allows us to control construction, but makes it more difficult for users to
define new context types (though it's not clear if this would be a common concern).

I have also made a separate construction method for having a default deadline
NewDefaultContext().  However it's worth keeping in mind that creating deadlines
leaks resources for the duration of the deadline unless the corresponding context
is canceled.

Change-Id: I2fa4c1d64ff6207286ffff828ed1a7aa1588afc3