veyron2/security: Improve signature scheme.
Prior to this commit, when a "purpose" was provided to Sign
operations, the message to sign was:
hash(hash(message), message+purpose)

This commit changes it to:
hash(hash(message), hash(purpose))

Motivations:
(1) More predictable memory usage
    (appending hash(purpose) to hash(message) is an operation
    over two fixed-with slices of small sizes (upto 64 bytes),
    while the old scheme used append(message, purpose...), which
    consists of two arbitrary sized slices
(2) Adhere to a simple pattern for creating a "content hash" of
    multi-field structures. This pattern will be used in upcoming
    implementations of the security model APIs that we are moving towards.

For backward compatibility (specifically, to keep any existing
veyron identities usable), special cased the "nil" purpose for now.
Once the new API is completely in place, there should be no usage
of the "nil" purpose and this hack will be removed.

Change-Id: Ibd24cbce5c20086ed98336094d8cce1cf41f1b93