veyron2/security: Message digest for the Caveat type.
Motivation:
The Certificate implementation (and upcoming third-party caveat
implementation) requires only a message digest of the Caveat type.
By using that instead of Caveat.ValidatorVOM directly, we allow
for future changes to the Caveat type (more fields being added,
say we do away with VOM for example) without breaking existing
Certificates or having to comb-through all usage of ValidatorVOM
and deciding if the new fields in the Caveat struct should be
incorporated.

Change-Id: Ia6165df6b5b861da1d61c1e76bf66991d78a19b3