veyron/services/mgmt/pprof: Add pprof proxy interface.
This change makes all the pprof data available via RPC such that it is
possible to run pprof against a veyron server.

Since the pprof tool doesn't speak the veyron protocol, we still need a
proxy on the client side to convert between the veyron interface and HTTP.

I'm working on a separate change that will add a command-line tool to
make interfacing with pprof, and the debug server in general, easier.

Change-Id: I9b315af2759175d267fea9765bfcba7cafb2e838