veyron2/security: VDL specification for Blessings.
With this change:
- Other VDL files can refer to a VDL type for blessings (MarshaledBlessings)
- Can convert from this VDL type to the interface type with a factory function

This will allow all usages of VOM-decoded blessings to decode into a
concrete type (as opposed to an interface).

Furthermore, once VDL-config-file support comes along for converting between
a VDL type and an in-memory representation (using a factory function),
users will not have to invoke security.NewBlessings manually, the framework
will take care of it.

Change-Id: Iea44d0e96dbe6dbe53fcf31a05c2e8c48e03d337