"veyron2/security": Split MintDischarge
The MintDischarge method currently both determines whether a
third-party caveat is discharageable (i.e. restrictions embedded
in it are valid), and also mints a discharge for the caveat.
This makes it difficult to expose the MintDischarge method via
an "agent" principal as the agent may not be able to VOM-decode
caveats.

This CL splits the MintDischarge functionality into two methods
- A method on ThirdPartyCaveats that checks whether the caveat is
  discharageable for a context.
- and the other on Principal that mints a Discharge for a given ID
  and a set of discharge caveats.

Change-Id: Ifdeaa730bce1d792b4e96c4363edc7bdfa48221b