veyron2/vom2/vomtest: Add vomtest tool to generate vom test data.
The concept is to generate vom test data, so that adding new test
cases is easy, and also to make it easy to update the tests if
package paths are changed, or we change the encoding format.

We start with the vomdata file.  This is the source file for VOM
test data, in the VDL config file format.  The purpose of this
file is to make it easy to add VOM test cases for each generated
language.

In order to add new test cases, simply add new values to the list
below.  Test types are defined in the vomtype.vdl file in this
directory; they can't appear in this file since they're not
allowed in VDL config files.

To re-generate the tests:
1) veyron go install veyron2/vdl/vdl veyron2/vom2/vomtest
2) veyron run vomtestgen
3) veyron run vdl generate veyron2/vom2/...

Running "vomtestgen" against this file produces the
vomdata.vdl file, and running "vdl generate" against the
resulting package produces the VOM test data in each generated
language.

In addition to the vomtest tool, this CL also includes various
features and bugfixes to make things work.

Change-Id: Ieff49843a15f1ff7489c86fbe93496bd85ec11f9