veyron2/vdl: Change vdl to use VDLPATH, rather than GOPATH.
Since vdl files are agnostic to the generated language, they
shouldn't depend on GOPATH; the fact that they did before was a
historical accident.  This change is a prerequisite for moving
vdl files out of the go repos.

Also added an additional help topic for "vdlpath" in the vdl
tool, so that users can get complete documentation.

Change-Id: I3773571431019b21b1c24a8312f81a49c724ef68