veyron/security: Provide a common implementation of the BlessingStore and BlessingRoots interfaces.
This commit moves the BlessingStore and BlessingRoots implementation
from veyron/runtimes/google/rt to veyron/security and provides
simple factory functions for creating a Principal implementation
that uses these (canonical) BlessingStore and BlessingRoots implementations.

As a result, a couple of silly BlessingStore and BlessingRoot implementations
that were used only in tests are now deleted.

Change-Id: Icb0cfe98351b82a59951e1cd1b795007de92abbd