veyron2/vlog: Add a clarification around ConfigureLogger for behavior that's extremely confusing: some options only work when set immediately after the logger is created, but before it's used.
For example, trying to set LogDir after the logger has already been used to
log something will be silently ignored. This is especially hard to debug
with the global Log singleton, which may get initialized by an init() function
imported through runtime or testutil -- and hence, any further call to
vlog.ConfigureLogger is a no-op as far as setting LogToStderr or LogDir is
concerned.

Change-Id: I78efbe5283614ea5e15f7a08dbdb2c91a3e7c55e