"veyron2": RuntimePrincipal option
This CL adds an option (ROpt) to specify a Principal that should be
used to initialize a runtime. This option would be useful in a few places,
e.g., in WSPR while initializing runtimes for per-app Controllers based on
the principal created for the app, in the identity service to initialize the
runtime with a specific auditing principal, etc.

Change-Id: I1537b021b5107efdbe1cc5ebf7baeaa2bdc4eaf0