veyron2/vdl: Change vdl package syntax to be identical to go.
The previous implementation of package specification for the vdl
tool was similar to the go tool, but not identical.

E.g. the go tool only interprets a path as a filesystem dir if
it's absolute, or starts with a . or .. element; the old vdl tool
would check if the path is absolute, but would consider any path
that starts or ends with a . as a filesystem dir.

Also, the go tool interprets wildcards "..." as a regexp ".*",
with a special-case for trailing x/...; the old vdl tool only
allowed the x/... special-case.

The new implementation now behaves identically to the go tool.
As a result I've also changed the wildcard strategy to use an
explicit walk, rather than recursive calls, in an attempt to
simplify the logic.

I've also added some preliminary tests, and moved the test_arith
and test_base packages under a testdata directory, so that it
doesn't clutter the godoc.

Change-Id: I28c81c7665c4bc2090d0acfe063e40dcec200802