veyron2/vdl: Add vdl.config file support.
The "vdl.config" file is an optional file in each vdl package
source directory.  The format is described by the vdl.Config type
in the standard "vdl" package; it uses the generic vdl config
file support to parse the actual config file, which is a variant
of the regular vdl syntax.  The "vdl.config" file implicitly
imports the standard "vdl" package, so that no imports are
necessary in these files.

The standard "vdl" package lives in a VDLROOT directory; just
like Go has GOROOT and GOPATH, vdl has VDLROOT and VDLPATH.  If
VDLROOT isn't specified, we construct it via VEYRON_ROOT; it is
an error if neither VDLROOT nor VEYRON_ROOT is set.  Other
standard vdl packages (like time) will be added under VDLROOT.

This CL doesn't add any actually useful options to "vdl.config";
it gets everything set up so that it should be easy to add such
options moving forward.

Change-Id: I6184fa829b7adb409ae06a5cb208fca803eeee04