veyron2/ipc/model.go: Add the new Glob interfaces.
This change is the 1st of a series of changes that add a simple
interface for objects to participate in the namespace.

Each object will have the option of implementing the full
mounttable.Globbable interface, or the much simpler VChildrenGlobber
interface where each object only has to enumerate its immediate
children.

The subsequent changes will update the generated vdl.go files in other
repositories and add the new interface to all the invoker implementations.

Then we'll modify the ipc.Server to implement Glob using the new
interface, and finally, we'll convert most of the existing objects to
use VChildrenGlobber.

Change-Id: I68acd6437c3d083609fa4debf5829faf085f8681