veyron2/naming: Move veyron2/services/mounttable/types here.
We have two copies of the types:
  MountedServer
  MountEntry

Before this change, they lived in the packages veyron2/naming and
veyron2/services/mounttable/types.

After this change, they both live in the package veyron2/naming.
I've added a VDL prefix to the vdl-based ones, as a temporary
measure.

Once vdl supports a standard time package, we can consolidate
these types back into one representation.

The motivation for this change is an upcoming change to our VDL
generated server stubs, which will need these types to be defined
in a standard location so that the Globbable interface can be
defined with typesafe stubs.

Change-Id: I8ffb20e095afbcee06f2c8e83750b1c28aee7caa