vom: Make go vom output consistent with javascript output
The largest change here is making the type encoder output in depth first order.  Note that the type ids are sent out of order so that they can be allocated in advance to prevent cycles. This can be changed if it is not desired, but this makes the code simpler.

Change-Id: I4dab87392a296b7af8563dfb82ba4c3b961eadc6