veyron2/verror2: Use defaults for language and programme name if not in context
Previously, I assumed that most code would have a context, even if only a default one,
and that if a context could be found, it would contain valid language and programme names.

However, many programmes don't set up even a default context (or runtime),
and sometimes when a context is present, it is not one derived from a
standard runtime, and so the language and programme names might be absent.

This change forces defaults for these two names when they are otherwise absent:
en-US for the language, and filepath.Basename(os.Args[0]) for the programme.

These defaults are set currently only when an attempt is make to use a context.
If ExplicitMake or ExplcitConvert are used, the code continues to do exactly
what the client requested, even if some of the names are empty.

Change-Id: Id44d14ea8e3e8200ecf92fc4076ef3f7faf1c89a