veyron.io/veyron/veyron2/i18n: Support colon-format strings in message catalogues.
In i18n message catalogues, the string {3} would be replaced by
parameter 3 (and similarly for other parameters).

With this CL, in addition:
        {:3} causes the parameter to be prefixed with ": "
        if the paramater is not the empty string.

        {3:} causes the parameter to be followed by ":"
        if the paramater is not the empty string.

        {:3:} causes the parameter to be prefixed with ": " and
        followed by ":" if the paramater is not the empty string.

So the format string
        "{1} foo {2:} bar{:3} wombat{:4:} numbat"
would yield
        "1st foo 2nd: bar: 3rd wombat: 4th: numbat"
if given the parameters "1st", "2nd", "3rd", "4th"
and would yield
         " foo  bar wombat numbat"
if given four empty-string parameters.

This will allow us to write verror2 error formats like this:
	"{1:}{2:} Does not exist{:_}"
to get error messages like this:
	stored:"wombat".Get: Does not exist: directory_name

This isn't quite right for right-to-left languages,
because of the asymmetry in spacing behaviour of {:1} vs. {1:},
but should still be siginificantly better than previously
If there is demand, we can add other options.

Change-Id: I1d92badd53aaefe97b6a535b410578fe37d6f9e3