veyron2/vdl: Prepare vdl types for vom2 integration.
This is a collection of many changes that are pre-requisites for
the vom2 integration:

  1) Change oneof to be labeled, like structs.  This is hard to
     change piecemeal, so I just changed everything in one shot,
     including the vdl type / value system, reflection, value
     conversion, the vdl parser, compiler and code generators,
     etc.  The tests under veyron2/vdl/... all pass, but I've
     left some TODOs to go back and clean things up, and also for
     java and javascript codegen.

  2) Move optional out of experimental.  I'm pretty sure some
     stuff won't work and isn't fully tested yet; I'll fix that
     up after this CL.

  3) Misc other changes - e.g. fixed some bugs, minor changes to
     enum codegen, etc.

Note that the new codegen for oneof is easier to use, and also
allows us to capture enough information via reflection such that
we can always create the Go objects corresponding to a oneof,
without any type registration.  Here's the strategy:

   // type Foo oneof{A bool;B string;C int32}
   type (
    Foo interface {
      Index() int
      Name() string
      __DescribeOneOf(__FooDesc)
    }
    FooA struct{ Value bool }
    FooB struct{ Value string }
    FooC struct{ Value int32 }

    __FooDesc struct {
      Foo
      A FooA
      B FooB
      C FooC
    }
   )

The trick is in the Foo{A,B,C} structs, which are the concrete
field structs for oneof - the type tells us both the label as
well as the type of the underlying field value.  We include these
in __FooDesc so that we can easily retrieve this info via
reflection, by looking for the __DescribeOneOf metadata function.

Change-Id: I92f04db58cb6ede743d6d9f800cb4ed16072f237