veyron/services/mgmt: Implement optional packages
This change implements optional packages in the mgmt services. The
binary repository now saves the media type and encoding of the objects,
e.g type=application/x-tar,encoding=gzip, or type=application/zip.

The node manager uses this information to download and install the
packages requested in the application envelopes.

TODO: Test the node manager implementation. This will require either using the
real binaryd, or significantly improving the mock binaryd.

Change-Id: Idc7d0ee7663ccd1816a271289de8f61403d8dcce