veyron2/services/security/access: Hack for the auto-generated Signature method.
After http://go/vcl/7090, a service using the TaggedACLAuthorizer would
not allow anyone to call the Signature method (generated by the VDL
compiler) on the service, since there is no access tag on that method
and thus no ACL associated with it.

(Prior to http://go/vcl/7090, only Admins could call the method anyway,
so it isn't as if that policy was ideal :)

This commit introduces a hack to expose the Signature method to
everyone. This hack should be removed before release and details
on why is in the comments.

Change-Id: I2857630eff5f3e17d77634f79f1af9bdb3992e78