veyron2/vdl: Fix package path for standard vdl packages.
This CL fixes our annoying issue go/veyron:bug/495, where the
package path for the standard "vdl" package would end up as
"veyron.io/veyron/veyron2/vdl/vdlroot/src/vdl" rather than "vdl",
but only sometimes.

The problem is that prior to this change, the vdl build package
had two ways to set the package path for any given package.

If the package was being built via a directory, either explicitly
or via ... wildcards, we'd attempt to deduce the package path
given the SrcDirs.  That would always result in the correct "vdl"
package path.

But if the package was being built via an import path, we tried
to be "smart" and skip deducing the package path, since the
import path was already provided.  And if the import path
happened to be "veyron.io/veyron/veyron2/vdl/vdlroot/src/vdl",
we'd assume that was the correct package path, when in fact it's
completely wrong.

The fix is to remove the "smart" case and always deduce the
package path given the SrcDirs.

Note that this was especially hard to track down, since the
"veyron go" command translates all packages on the command line
into package paths, even if wildcards or directory paths were
explicitly specified.  E.g. say we ran this with the old code:
  veyron go install ./veyron2/vdl/vdlroot/src/vdl

This is clearly a dir path, since it starts with "./", but the
veyron tool takes the dir path and computes the transitive Go
package dependencies, and ends up invoking the vdl tool as:
  vdl generate -lang=go veyron.io/veyron/veyron2/vdl/vdlroot/src/vdl

Note that the dir path has been transformed into a package path,
which triggered our buggy behavior.

Change-Id: I9ab6911bd228fa6ece4d8838a71e6e5a314fd24b