veyron2/{vdl,vom2}: Add vom debugging tool.
[Since there are vdl compiler changes, to build run this:
  veyron go -novdl install veyron.io/...
]

Also minor changes to Go codegen for VDL enum, so that all
generated enums implement the flag.Value interface.  This makes
it convenient to use the vdl tool to generate enums used as
flags.

We current support two commands: decode and dump

=================================================================
Vom Decode

Decode decodes data encoded in the vom format.  If no arguments
are provided, decode reads the data from stdin, otherwise the
argument is the data.

By default the data is assumed to be represented in hex, with all
whitespace anywhere in the data ignored.  Use the -data flag to
specify other data representations.

Usage:
   vom decode [flags] [data]

[data] is the data to decode; if not specified, reads from stdin

The vom decode flags are:
 -data=Hex
   Data representation, one of [Hex Binary]
=================================================================
Vom Dump

Dump dumps data encoded in the vom format, generating formatted
output describing each portion of the encoding.  If no arguments
are provided, dump reads the data from stdin, otherwise the
argument is the data.

By default the data is assumed to be represented in hex, with all
whitespace anywhere in the data ignored.  Use the -data flag to
specify other data representations.

Calling "vom dump" with no flags and no arguments combines the
default stdin mode with the default hex mode.  This default mode
is special; certain non-hex characters may be input to represent
commands:
  . (period)    Calls Dumper.Status to get the current decoding status.
  ; (semicolon) Calls Dumper.Flush to flush output and start a new message.

This lets you cut-and-paste hex strings into your terminal, and
use the commands to trigger status or flush calls; i.e. a
rudimentary debugging UI.

See veyron.io/veyron/veyron2/vom2.Dumper for details on the dump
output.

Usage:
   vom dump [flags] [data]

[data] is the data to dump; if not specified, reads from stdin

The vom dump flags are:
 -data=Hex
   Data representation, one of [Hex Binary]

Change-Id: Ia314492ab84fa5f54a99260fa843448ec9b43c93